Added pathfinder api fix & some other service fixes

+++ src/main/java/com/blazarbit/neo4j/model/dto/PathfinderDTO.java
@@ -19,5 +19,5 @@ public class PathfinderDTO {
     private String address;
     private String feeToken;
     private Long feeAmount;
-    private Long destinationTokenAmount;
+    private Double destinationTokenAmount;
 }
